requirements.txt
streamlit==1.31.0
pymupdf==1.23.8
pdfplumber==0.10.3
python-docx==1.1.0
openai==1.50.0
httpx>=0.23.0,<0.28.0
fpdf==1.7.2
python-dotenv==1.0.1
bcrypt==4.1.2
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
tenacity==8.2.3
pydantic==2.6.1
chromadb==0.4.22
